requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
boto3==1.35.98
botocore==1.35.98
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
cssselect==1.2.0
cssutils==2.11.1
Django==5.1.4
django-allauth==65.3.0
django-cors-headers==4.6.0
django-environ==0.11.2
django-otp==1.5.4
django-redis==5.4.0
django-soft-delete==1.0.16
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
drf-yasg==1.21.8
email_validator==2.2.0
frozenlist==1.5.0
idna==3.10
inflection==0.5.1
jmespath==1.0.1
lxml==5.3.0
more-itertools==10.5.0
multidict==6.1.0
packaging==24.2
pillow==11.0.0
premailer==3.10.0
propcache==0.2.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-http-client==3.3.7
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
s3transfer==0.10.4
secure-smtplib==0.1.1
sendgrid==6.11.0
six==1.17.0
sqlparse==0.5.2
starkbank-ecdsa==2.2.0
twilio==9.3.8
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
yagmail==0.15.293
yarl==1.18.3
